Add files via upload
In this the notebook file named as "EMP_SALARY MODIFIED.ipynb" has the code for preprocessing techniques,comparision of models to train the data ,and ultimately choosing the best out of them to train the model. This file also includes the deployment part with the inclusion of "ngrok". For deployment without the use one can directly involve the streamit or else the file named as "app1.py" can be used to deploy the model (as long as the pkl file is in the same directory).